+++ modules/core/src-js/PlatformCompat.scala
@@ -2,4 +2,7 @@ package weaver
 
 private[weaver] object PlatformCompat {
   val platform: Platform = Platform.JS
+
+  def getClassLoader(clazz: java.lang.Class[_]): ClassLoader =
+    new ClassLoader() {}
 }
