+++ modules/core/src-jvm/PlatformCompat.scala
@@ -2,4 +2,7 @@ package weaver
 
 private[weaver] object PlatformCompat {
   val platform: Platform = Platform.JVM
+
+  def getClassLoader(clazz: java.lang.Class[_]): ClassLoader =
+    clazz.getClassLoader()
 }
